C Server : Add iOS build PR validation and fix platform build

+++ servers/c/scripts/build_ios.sh
@@ -1,17 +1,18 @@
 #!/bin/bash -e
 
 function usage() {
-    echo "Usage: $0 <edition: enterprise | community> <version> [build num]"
+    echo "Usage: $0 <binary type: all | simulator | device> <edition: enterprise | community> <version> [build num]"
     exit 1
 }
 
-if [ "$#" -ne 2 ]; then
+if [ "$#" -ne 3 ]; then
     usage
 fi
 
-EDITION=${1}
-VERSION=${2}
-BLD_NUM=${3}
+BIN_TYPE=${1}
+EDITION=${2}
+VERSION=${3}
+BLD_NUM=${4}
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 ASSETS_DIR="${SCRIPT_DIR}/../assets"
@@ -58,15 +59,21 @@ pushd vendor > /dev/null
 sh cmake.sh
 popd > /dev/null
 
-# Build
-set -o pipefail # Get xcpretty to report failures
-xcodebuild -scheme TestServer -sdk iphonesimulator -configuration Release -derivedDataPath "${BUILD_SIMULATOR_DIR}" | xcpretty
-xcodebuild -scheme TestServer -sdk iphoneos -configuration Release -derivedDataPath "${BUILD_DEVICE_DIR}" -allowProvisioningUpdates | xcpretty
-
-# Copy artifacts
+# Build and Copy Artifact
 rm -rf "${BUILD_BIN_DIR}"
 mkdir -p "${BUILD_BIN_DIR}"
-cp -r "$BUILD_SIMULATOR_DIR/Build/Products/Release-iphonesimulator/TestServer.app" "${BUILD_BIN_DIR}/TestServer-Simulator.app"
-cp -r "$BUILD_DEVICE_DIR/Build/Products/Release-iphoneos/TestServer.app" "${BUILD_BIN_DIR}"
+
+set -o pipefail # Get xcpretty to report failures
+if [ ${BIN_TYPE} != "device" ]
+then
+    xcodebuild -scheme TestServer -sdk iphonesimulator -configuration Release -derivedDataPath "${BUILD_SIMULATOR_DIR}" | xcpretty
+    cp -r "$BUILD_SIMULATOR_DIR/Build/Products/Release-iphonesimulator/TestServer.app" "${BUILD_BIN_DIR}/TestServer-Simulator.app"
+fi
+
+if [ ${BIN_TYPE} != "simulator" ]
+then
+    xcodebuild -scheme TestServer -sdk iphoneos -configuration Release -derivedDataPath "${BUILD_DEVICE_DIR}" -allowProvisioningUpdates | xcpretty  
+    cp -r "$BUILD_DEVICE_DIR/Build/Products/Release-iphoneos/TestServer.app" "${BUILD_BIN_DIR}"
+fi
 
 popd > /dev/null
